Add Graveyard mob stats
Experience values for Fallen and Spectre have been entered manually.

The formula as it is currently gives a lot of emphasis to dodge and def,
which magic bypasses easily. Hence, the Spectre, which was designed
to have a vulnerability to magic, was given a great deal higher exp reward
than it should have. The Fallen suffered the reverse, and was somewhat
underrated.

Exp calculation in general still needs modification with the addition of
magic, but it still works fairly well with the other additions.