Give the Scythe +100 while removing the shielding.
Most players don't have a shield by the time they get the Scythe other than the leather shield which makes little difference in saving a player (and isn't intended to either).
Most players never keep the scythe for great lengths of time because the Short Sword soon replaces it. (It is a rarely seen tool used in game, from observation.)
Many players bypass the tool altogether and are led to the Short Sword by helpful higher level players. And that is not a reflection on the lack of shielding.
This makes it as strong as the Short Sword (too easily attainable these days) but still at a disadvantage.

Ref. https://forums.themanaworld.org/viewtopic.php?p=75932#p75932