Fixing a bug where you could not talk to NPCs.
Problem: When you try to talk to some NPCs and these NPC don't speak
to you, then after that you can't talk to any other NPC until the
client is restarted.

Solution: The responsibilities of the rules NPC are moved to Ian.
The well and the tree are talking to everyone now.