fixed minor bugs in orum/waric
  * instead of "Intensity out of range" you die now.

  * spelling, only typos, no grammar corrected, since their grammar is weird.

  * header added to orum file.

Not done:
The bound checking is not corrected yet:
Instead of :

if (@offsetOne > 12) set @offsetOne, 1;
if (@offsetOne < 1) set @offsetOne, 12;

you should do:

if (@offsetOne > 12) set @offsetOne, @offsetOne - 12;
if (@offsetOne < 1) set @offsetOne, @offsetOne + 12;

This can handle with steps bigger than 1.

Collision layer of the bed in the empty cave must be fixed.

Should there be hints how to setup the torches anyway?