setup Miriam's timing for quest
Ok on my local testserver i could do it with a lvl 80 char and lots
of healing potions in 200 seconds.
(I was two times wrong in taking the right path,
so probably you can be 10 seconds faster, so fastest i guess 190 secs.)

Now you must be faster than 210 seconds to succeed.

So there are 20 seconds of reserve, which are about 10% of the whole time.
That should be enough for even bad connections.