Cindy: Spawning one more Yeti at each wave
With the last bug fix the quest got a bit easier. This commit
makes it a bit more difficult again. If the players manage to
kill all Yetis before the round timer triggers the next wave,
then the maximum number of Yetis alive is the same as in the past.

For the benefit of the hero's companions Cindy now announces the 0th
wave publically.