.:: VALENTINE 2020 ::.
Adds three new commands:
@givelvlplz and @valentine2020 are the same command - They are for users
@moubootaurs is for staff.

User command: Allows them to insert the needed credentials on ML. It is
already authenticated inside the game, so they only need to inform ML data.

Staff command: Obtain the listing of users which used the previous command
to **manually** insert on Moubootaur Legends special merge page.

As soon as the staff inserts the command, if there are no errors, the user
will be able to claim their levels from TMW at Moubootaur Legends. Temporaly.
If this was rEvolt, we could just try a SQL-transaction...

The tables are never cleared. I'm also saving at Account Register to don't
pollute char database. (We might want a clear vars later).
This also allows us to track down the responsible for each register in case
of issues.