Release Assassin Set (#550)
Add assassin set as a brawnling set to good ol' Iilia.

Now Luvia collected an useless unique gear!
Your chances to get the Heart Of Isis just went down from 25% to 20%!
Who'll be the first unlucky player to obtain the Assassin Shirt?!

PS. Cannot be equipped with weapons. But can be equipped with shields.
It does not boost attack strength - focus if you want to deal any serious damage.
Also, GMs are able to equip it even while armed, but they'll suffer an agi penalty.
This got spammy and buggy, so it is for GMs only from time begin.

PPS. Brawling is a non-canon combat style. Do not expect to be able to actually kill stuff with it.
Oh, but it probably will be fun. Tell me if it needs any other serious tweaks.