Add to @numa Quest Debug options to change Candor Battle difficulty. (#540)
Now features Easy (1800) Normal (2200) Hard (3300) and Expert (4500)
Do not use Expert unless you know what you are doing. These settings are meant
for events only; The only way to revert is by Numa or a server restart.
The rewards are always the same regardless of difficulty setting.
There's no option to change minimum players to begin a Candor Fight, because I
didn't deem that to be necessary; Not like there would be an event with < 5 people
online.

Evolved has a superior framework, this will not be ported back.