Added Dragon Shield to Chronos - 4 BM
It's been requested a few times and there is no way to get it in the game, currently. If we add a quest, we could remove from the store, but since there's already big quest items there, I see no need to take it out unless the others go too.

Approved-by: HoraK-FDF <horak-fdf@web.de>