Ice Gladius Buff
It's rarely used when compared to the Setzer and is of the same required level(60) to do the quest to obtain them. The Setzer, being a knife is faster and faster yet from a small bonus and does 100 dmg. The Ice Gladius does 130 dmg, is slower and gives +1 luck. I added +5% crit defense to buff it a little. The rock knife(lvl 70) still gives more bonuses with +3 vit and does 110 dmg, but is also fast as it is a knife. I would think this may entice players to use the Gladius more often, while not replacing the Setzer or Rock Knife.

Approved-by: HoraK-FDF <horak-fdf@web.de>