Adjust Warlord Plate & Helmet and Leather Gloves drop rates
This is a conspiracy by Small Nerf!

j/k; the armors and helmets are dropping way too often and so nobody
crafts them. The gloves I have 50 in my storage, but they are only
nerfed down to 1:2000 as they're uncraftable otherwise.

There has been a Warlord Plate listed @ ManaMarket for weeks now at 150k gp
and nobody is purchasing it. This price barely covers the price of making
the required ingots and what Nicholas's apprentice charges to make the armor.