Migrated the "furquest" (Agostine, Lora Tay) quest line to the quest log
This required a sizable refactor to update and clarify most of the scripts, making it easier to read. My goal is to preserve the original quest line in its entirety, as the dialog is perfectly fine, but to also add it to the player's quest log.

I've broken the single monolithic quest into three distinct parts:

* Winter Gloves
* Furry Boots
* Fine Dress

Each of these will be a distinct entry into the quest log. Details like names and wording can be tweaked later, but I feel this has the best feel from a player perspective. Other elements like dialog changes and adjusting the RNG logic should be addressed in another MR.

Addendum: This will need [this MR](https://git.themanaworld.org/tmw/clientdata/-/merge_requests/694) for clientdata as well.

Approved-by: Hello TMW <hello@themanaworld.org>
Reviewed-by: Hello TMW <hello@themanaworld.org>